# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a code challenge that manages some bank routines.",
        "title": "Digibank API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Caique Nunes",
            "url": "https://www.linkedin.com/in/caique-nunes/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/accounts": {
            "post": {
                "description": "This endpoint creates a new account and returns the \"document_number\" (string) representing the accountID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "This endpoint creates a new account",
                "parameters": [
                    {
                        "description": "Document number to be inserted into account to be created",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{\r\n  \"document_number\": \"12345678900\"\r\n}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountPresenterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutputDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutputDto"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{accountID}": {
            "get": {
                "description": "This endpoint receives an accountID and returns the respective account details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "This endpoint retrieves a specific account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "accountID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountPresenterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutputDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutputDto"
                        }
                    }
                }
            }
        },
        "/api/v1/transactions": {
            "post": {
                "description": "This endpoint receives a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "This endpoint receives a transaction",
                "parameters": [
                    {
                        "description": "Body of a transaction",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{\r\n\"account_id\": 1,\r\n\"operation_type_id\": 4,\r\n\"amount\": 123.45\r\n}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutputDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutputDto"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "This description endpoint is a health check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HealthCheck"
                ],
                "summary": "This summary endpoint is a health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AccountPresenterResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "document_number": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorOutputDto": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}